[wip][ci skip] Remove ingressroute
