+++ config/rbac/role.yaml
@@ -33,20 +33,6 @@ rules:
   - update
   - watch
 - apiGroups:
-  - contour.heptio.com
-  resources:
-  - ingressroutes
-  verbs:
-  - get
-  - list
-  - watch
-- apiGroups:
-  - contour.heptio.com
-  resources:
-  - ingressroutes/status
-  verbs:
-  - get
-- apiGroups:
   - externaldns.k8s.io
   resources:
   - dnsendpoints
